Accept optional type guard in DynamicConfig getter
The behavior prior to this change was only type-safe for simple
primitive types. Complex objects or Arrays of objects would be
dangerously type-cast.

This change addresses the issue by accepting an optional type guard
function as a third parameter. If it is not supplied, the behavior
remains unchanged from prior releases. If it is supplied, the property
access is checked against the type guard prior to returning. If the type
guard resolves to false, then the default value is returned instead.

In cases where a null or undefined default value is passed and type
guard is supplied for a non-nullable type in strict mode, the type
should resolve to `T | null` or `T | undefined`. Similarly, wider types
in the default will resolve to the wider type for the end value.
Narrower types will remain as the wider type `T`.

This allows for type-safe access of dynamic values (assuming the
type guard function is well-written).