[release] 4.49.0 - Introduces aggregated count checks for non exposed checks (#454)
In order to optimize payload size, we need to know which entities are
checked by any sdk. exposures work for this, but if you disable exposure
logging, we need a way to know you are accessing an entity from a given
sdk key. This adds those counters.

Internal updates to remove RN types from dev dependencies





>Included In This Release
>- 4811bf1daa6665c458ffcf6fa4f3c19de2a86c1d tore-statsig
>   - Remove rn dev dep types (#453)
>- 8ef133d1af547e396c206bf03da4053e6e52c15a sroyal-statsig
>   - Event For Non Exposed Checks (#452)

---------

Co-authored-by: statsig-kong[bot] <statsig-kong[bot]@users.noreply.github.com>